add transit state logic to identiy wallet
